move fuzzer code to _test files
