+++ pkg/apis/messaging/v1beta1/roundtrip_test.go
@@ -19,13 +19,68 @@ package v1beta1
 import (
 	"testing"
 
+	fuzz "github.com/google/gofuzz"
+	"k8s.io/apimachinery/pkg/runtime/serializer"
+	"knative.dev/eventing/pkg/apis/duck/v1/test"
+	"knative.dev/eventing/pkg/apis/messaging"
+
 	"k8s.io/apimachinery/pkg/api/apitesting/fuzzer"
 	"k8s.io/apimachinery/pkg/runtime"
 	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
-	duckerfuzzer "knative.dev/eventing/pkg/apis/duck/v1"
 	v1 "knative.dev/eventing/pkg/apis/messaging/v1"
 	pkgfuzzer "knative.dev/pkg/apis/testing/fuzzer"
 	"knative.dev/pkg/apis/testing/roundtrip"
+)
+
+// FuzzerFuncs includes fuzzing funcs for knative.dev/messaging v1 types
+//
+// For other examples see
+// https://github.com/kubernetes/apimachinery/blob/master/pkg/apis/meta/fuzzer/fuzzer.go
+var FuzzerFuncs = fuzzer.MergeFuzzerFuncs(
+	func(codecs serializer.CodecFactory) []interface{} {
+		return []interface{}{
+			func(ch *v1.Channel, c fuzz.Continue) {
+				c.FuzzNoCustom(ch) // fuzz the Channel
+				if ch != nil {
+					if ch.Annotations == nil {
+						ch.Annotations = make(map[string]string)
+					}
+					ch.Annotations[messaging.SubscribableDuckVersionAnnotation] = "v1"
+				}
+				// Clear the random fuzzed condition
+				ch.Status.SetConditions(nil)
+
+				// Fuzz the known conditions except their type value
+				ch.Status.InitializeConditions()
+				pkgfuzzer.FuzzConditions(&ch.Status, c)
+			},
+			func(imc *v1.InMemoryChannel, c fuzz.Continue) {
+				c.FuzzNoCustom(imc) // fuzz the InMemoryChannel
+				if imc != nil {
+					if imc.Annotations == nil {
+						imc.Annotations = make(map[string]string)
+					}
+					imc.Annotations[messaging.SubscribableDuckVersionAnnotation] = "v1"
+				}
+				// Clear the random fuzzed condition
+				imc.Status.SetConditions(nil)
+
+				// Fuzz the known conditions except their type value
+				imc.Status.InitializeConditions()
+				pkgfuzzer.FuzzConditions(&imc.Status, c)
+			},
+			func(s *v1.SubscriptionStatus, c fuzz.Continue) {
+				c.FuzzNoCustom(s) // fuzz the status object
+
+				// Clear the random fuzzed condition
+				s.Status.SetConditions(nil)
+
+				// Fuzz the known conditions except their type value
+				s.InitializeConditions()
+				pkgfuzzer.FuzzConditions(&s.Status, c)
+			},
+		}
+	},
 )
 
 func TestMessagingRoundTripTypesToJSON(t *testing.T) {
@@ -34,7 +89,7 @@ func TestMessagingRoundTripTypesToJSON(t *testing.T) {
 
 	fuzzerFuncs := fuzzer.MergeFuzzerFuncs(
 		pkgfuzzer.Funcs,
-		v1.FuzzerFuncs,
+		FuzzerFuncs,
 	)
 	roundtrip.ExternalTypesViaJSON(t, scheme, fuzzerFuncs)
 }
@@ -58,8 +113,8 @@ func TestMessagingRoundTripTypesToBetaHub(t *testing.T) {
 
 	fuzzerFuncs := fuzzer.MergeFuzzerFuncs(
 		pkgfuzzer.Funcs,
-		v1.FuzzerFuncs,
-		duckerfuzzer.FuzzerFuncs,
+		FuzzerFuncs,
+		test.FuzzerFuncs,
 	)
 
 	roundtrip.ExternalTypesViaHub(t, scheme, hubs, fuzzerFuncs)
